Making upload file PSR compliant
